+++ Package.swift
@@ -10,7 +10,7 @@ let package = Package(
   dependencies: [
     .Package(url: "https://github.com/Carthage/Commandant.git", majorVersion: 0, minor: 12),
     .Package(url: "https://github.com/drmohundro/SWXMLHash.git", majorVersion: 4),
-    .Package(url: "https://github.com/jpsim/Yams.git", majorVersion: 0, minor: 3),
+    .Package(url: "https://github.com/jpsim/Yams.git", majorVersion: 0, minor: 4),
     .Package(url: "https://github.com/norio-nomura/Clang_C.git", majorVersion: 1),
     .Package(url: "https://github.com/norio-nomura/SourceKit.git", majorVersion: 1),
   ],