+++ Package@swift-4.swift
@@ -8,9 +8,9 @@ let package = Package(
         .library(name: "SourceKittenFramework", targets: ["SourceKittenFramework"])
     ],
     dependencies: [
-        .package(url: "https://github.com/Carthage/Commandant.git", from: "0.12.0"),
+        .package(url: "https://github.com/Carthage/Commandant.git", .upToNextMinor(from: "0.12.0")),
         .package(url: "https://github.com/drmohundro/SWXMLHash.git", from: "4.2.0"),
-        .package(url: "https://github.com/jpsim/Yams.git", from: "0.3.6"),
+        .package(url: "https://github.com/jpsim/Yams.git", .upToNextMinor(from: "0.4.1") ),
         .package(url: "https://github.com/norio-nomura/Clang_C.git", from: "1.0.0"),
         .package(url: "https://github.com/norio-nomura/SourceKit.git", from: "1.0.0"),
     ],